+++ config/blog.php
@@ -1,5 +1,7 @@
 <?php
 
 return [
-
+	'routing' => [
+		'prefix' => 'blog'
+	]
 ];