+++ resources/views/bootstrap/blog/posts/create.blade.php
@@ -1,0 +1,3 @@
+<h1>
+    bootstrap create
+</h1>