+++ resources/views/bootstrap/blog/posts/index.blade.php
@@ -1,0 +1,3 @@
+<h1>
+    bootstrap index
+</h1>