+++ resources/views/tailwind/blog/posts/show.blade.php
@@ -1,0 +1,11 @@
+<h1>
+    tailwind show
+</h1>
+
+<div>
+    {{ $post->title }}
+</div>
+
+<div>
+    {{ $post->body }}
+</div>