+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/brucebrophy/laravel-blog/compare/v0.0.1...develop)
+- Adds Tailwind blog template
+- Updates create posts migration 
+- Adds .editorconfig 
+- Adds test for publishing command
+- Adds controller for posts
+- Adds routes for posts
+- Adds tests for posts 
+- Updates README.md
 
 ## [0.0.1] - 2020-07-04
 - Initial Skeleton of package is added